+++ modules/self-hosted/default.nix
@@ -31,8 +31,6 @@
     ./pyload.nix
 
     # Apps and utilities
-    ./codex.nix
-    ./agent-zero.nix
     ./cloakbrowser.nix
     ./n8n.nix
     ./bentopdf.nix